Improve DI
